Initial commit of some custom UI5 controls.
This initial commit contains the Box, FormLine, InfoBox and Page
controls which I built in order to simplify the generation of responsive
UI5 views with different content areas.
It also contains the Icon, StatusIndicator and Text controls which
enable the display of graphical information based on input from
bindings. (e.g. you can map an icon URL to a binding value or
indent/color text based on bound values.